move all the searches structures to new modules
